// src/app/modules/transportation/routing/model/student-transport-view.ts
import { RoutePointsView } from './student-route-points-view';
import { RoutePointsDataView } from './route-points-data-view';


/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 
/**
* StudentTransportView
*/
export interface StudentTransportView { 
    /**
    * Gets or sets
    */
    assigneeId?: string;
    /**
    * Gets or sets
    */
    transportScheduleInboundRouteId?: number;
    /**
    * Gets or sets
    */
    transportScheduleOutboundRoutesId?: number;
    /**
    * Gets or sets
    */
    requestedPickupAddress?: string;
    /**
    * Gets or sets
    */
    requestedDropoffAddress?: string;
    /**
    * Gets or sets
    */
    pickupPointId?: number;
    /**
    * Gets or sets
    */
    dropoffPointId?: number;
    /**
    * Gets or sets
    */
    pickupPoint?: RoutePointsView;
    /**
    * Gets or sets
    */
    dropoffPoint?: RoutePointsView;
    /**
    * Gets or sets
    */
    routePointDetails?: Array<RoutePointsDataView>;
}